Move drawing routines to auxillary sources
